Clean up and document collect-only test runner
